Move hindcast tests to hindcast folder
