+++ mhkit/tests/wave/io/hindcast/test_hindcast.py
@@ -22,7 +22,7 @@ import os
 
 
 testdir = dirname(abspath(__file__))
-datadir = normpath(join(testdir,'..','..','..','..','examples','data','wave'))
+datadir = normpath(join(testdir,'..','..','..','..','..','examples','data','wave'))
 
 
 class TestWPTOhindcast(unittest.TestCase):
